Make pushsource.helpers private
The helpers module is intended to include functions which assist
in implementing new sources, mainly to do with parsing arguments
from URLs.

I don't feel confident in this API, as the contents are fairly
arbitrary according to the sources which have been implemented so
far.  Let's make it private for now so we don't have to commit
to maintaining compatibility.