Fix crash on python2 if using get_partial with register_backend
For any registered backend, the Source class needs to know if
the backend accepts a 'url' argument, which is done by inspecting
the backend's signature.

However, on Python 2, this only works on "pure" python functions,
and not the output of functools.partial. This meant that if you
obtained a callable by get_partial and then gave it to
register_backend (which is the intended use), you'd get a crash
along the lines of:

  TypeError: arg is not a Python function

Let's avoid functools.partial and instead manually bind the function
we're returning, to improve Python 2 compatibility.