Introduce PushItem.with_checksums for calculating sums
As explained in doc string. Calculating checksums is potentially
expensive, and so it is not filled in implicitly by all pushsource
backends. The caller is able to control when sums are calculated

This utility method makes it simple to obtain checksums only when
needed, e.g. a Pub-style multithreaded "computing checksums" step
could be easily implemented via a threadpool executor by:

  items = executor.map(lambda i: i.with_checksums(), items)