tests: pin all dependency versions and enable dependabot
List specific versions of every dependency to be used in CI
in the test requirements files. This ensures that CI cannot break
due to backwards-incompatible changes in dependencies.

To ensure that these are kept up-to-date over time, dependabot
is enabled. It will create PRs automatically when new versions of
dependencies are available.