tests: drop dependabot for python2
This is a partial revert of the changes made in f2bb390196d50a to
introduce the usage of dependabot. After those changes were landed,
a few PRs showed that dependabot is not going to work properly for
the mixed python2 & python3 case, because it can't understand that
a new version of a dependency may only be available for py3, and
so shouldn't be added to the python2 requirements file.

PR #37 is an example of a broken update. more-itertools 5.0.0
was the last version supporting python2, but dependabot doesn't
understand this and attempts to add newer versions to
test-requirements-py2.txt.  There doesn't seem to be any way
to configure this.

Let's leave the py3 setup as it is, but revert the py2 setup back
to what it was before: test requirements don't pin a version, and
a constraints file is used to lock versions of specific deps as
needed.