cleanup: drop useless channel dumps code
Pushing channel dumps has not been done since 2019-03, and is not ever
intended to be used again. Hence, this code was obsolete before it was
written.

Rather than carrying support for this kind of push item which is not
ever expected to be used, let's clean it up before integrating "staged"
source into Pub.

Note a few converters which previously were tested only via channel
dumps had explicit tests added. The reason for doing this, rather than
simply removing the converters, is that the same field formats are also
used by content ISO push items - which are not currently supported, but
likely will be in future.