Add test-requirements.in to improve effectiveness of dependabot
Previously, test-requirements.in was not added to the repo.

This causes a problem: dependabot will not add entries to the
test-requirements.txt file for any new indirect dependencies
as they are introduced by upgrades to other packages. Since we
use hash checking mode, this causes CI to fail completely as the
resulting test-requirements.txt is not installable unless hashes
are present for *all* dependencies, both direct and indirect.

Adding test-requirements.in fixes this, because it causes
dependabot to run pip-compile in order to regenerate the
corresponding .txt, and pip-compile does support updating indirect
dependencies where needed, including hashes.

See also the (internal) mail https://url.corp.redhat.com/e051f38
for more context on the problem and solution.