staged: accept metadata for legacy unsupported content
The intended behavior for old unsupported content types such as
DOCKER is to log a warning and ignore discovered push items.

However, it is likely that the metadata file for the push also
has entries for those files. In that case, these must be able to
pass validation, otherwise we'll get an error rather than a
warning.

To make that work, add a loose schema for those unsupported types.
It does not validate any attributes on the metadata, as the goal
is simply to ignore these entries.