staged: fix overly strict erratum pkglist schema
The schema for the "pkglist" field within errata had set "name"
as a required field.

I have discovered some data in the wild for which that field is
missing, and after reviewing the code of pulp_rpm it seems that
data without this field is in fact able to be uploaded, so the
schema was too strict. Adjust so that 'name' (and 'short') are
both considered optional, with blank defaults.