Fix too strict handling of 'description'
This field in schema had been declared to have a minimum length of
at least one character (i.e. not an empty string), since it seems
that the description may as well be omitted if you're going to pass
an empty string.

However, in practice it seems that empty strings have genuinely
been used in the past here and files with empty descriptions (rather
than absent) do exist in the wild. It was not intended for the
schema to block any existing workflows, so let's loosen it up and
make it accept this data.