Fix missing reboot_suggested field from erratum pkglists
The reboot_suggested field is unusual in that it can appear
at two levels within an erratum: once at the top level, but also
within each entry under 'pkglist'.

The latter previously was incorrectly missing from the schemas and
model in this library, leading to a crash if an attempt was made
to push an erratum using the field in pkglist. This is important
because it's actually the field in pkglist which is consumed by
yum (and the usage of the other field is unclear - it's possible
it only exists for historical reasons).

Let's update the model and schema to accept reboot_suggested in
pkglist, update a few tests to ensure that this is covered, and
also add a warning about the quirkiness of this field to the docs.