Fix argument leaks in get_partial
Previously if you've prepared a source like this:

  src = Source.get_partial('somebackend', ...)
  inst1 = src(key1='value1')
  inst2 = src(key2='value2')

...then the second call would be passed both key1 and key2
incorrectly.

This happened due to some code which intended to merge two
dicts together, but it did this via an in-place update of
a dict used at every call rather than making a copy.
Fix it to do the merge safely.

I don't think this affects any current usage of the library,
but it would interfere with errata source being prepared in
RHELDST-5430.