Add a reset() feature to remove all custom backends
When a user of this library needs to customize backends, as is the
case in Pub, writing tests for that code can be cumbersome as the
tests would have to carefully undo any customizations made to ensure
that custom backends aren't left behind interfering with later tests.

Let's add a reset method so we can simply do "Source.reset()" during
test cleanup, to put everything back to the original state.

Refactoring connected with RHELDST-5431.