docs: fix missing doc for Source.__iter__
Special methods (__foo__) are not included in docs by default,
even if they have doc strings.

It makes sense to include this one, since the whole point of a
Source instance is to be iterated over. The purpose of the class
may be rather unclear if this can't be seen.