koji: use NSVCA as module push item name
Testing this code with Pub revealed a behavior difference: in Pub
code, we load modulemd files and use their NSVCA as the push item
name. In this library we were not doing that, and instead using
the modulemd filename as push item name.

Although it's a bit annoying that we have to parse the modulemd
files before we can return push items, overall the Pub behavior
seems better as the NSVCA is a much better identifier for what's
being pushed than a generic filename like modulemd.<arch>.txt.
So let's do the same here.

As a compromise, we still use the filename as a fallback when the
file isn't available. This is done since loading the file needs
koji's NFS volume available, which is true in production but won't
be the case in many development and testing scenarios.