Introduce separate class for modulemd sources
We had previously used ModuleMdPushItem to represent any
modulemd-related archives on a koji build. This included both built
modules such as modulemd.x86_64.txt and module source files such as
modulemd.src.txt.

We shouldn't really do that though, because they're very different
things. The document format is not the same, and the expected usage of
each type of push item is completely different (e.g. modulemd.src.txt
are not meant to be uploaded to yum repos).

Let's model them as separate types so the caller can process the two
types differently and won't accidentally try to handle a source modulemd
in the same way as a built modulemd.