Add preliminary support for container-related push items
This commit introduces support for two new push item types which can now
be located by errata or koji source:

- ContainerImagePushItem: a container image

- OperatorManifestPushItem: operator manifests bundle typically found
  alongside container images in koji

This is referred to as "preliminary support" because these classes
haven't defined any attributes yet (i.e. they only have the inherited
attributes from the base PushItem class). This is done to avoid the
change becoming too large and complicated.

In this commit, we focus only on how we locate these push items when
given an advisory; filling out the model with all needed attributes
will come later, and it's expected to be rather complicated to figure
out what's needed there.