docs: add an attribute index onto push items
Previously, in the docs for push item classes, inherited attributes
were not displayed. This is annoying since it means one needs to look at
the base class and subclass docs together to understand what's
available. In extreme cases such as ProductIdPushItem which don't have
any of their own attributes, it gives the impression of a completely
empty class.

Unfortunately none of the built-in sphinx autodoc features seem able to
solve this.  There is an :inherited-members: option, but if enabled it
copies the entire doc strings from the base classes without any
indication of what's inherited and what's not, which is far too
verbose.

So let's add a little extension ourselves which generates a list under
the class description.