tests: introduce baseline tests
For test cases which work simply like this:

- get a Source via URL
- compare all generated items against expected

...the prior approach was to write python code executing the above
steps.

A smarter way to do it is to put the (inputs, expected outputs) into
test data files which can be automatically rewritten. Let's do that
where feasible. Tests which are obsoleted by this have been deleted.

The primary motivation here is that the container push items will need
to have many fields added in upcoming commits. If there are tests
written in python doing equality checks between actual and expected push
items, those will fail every time a new field is introduced, and since
it's python code there is no easy way to automatically update those
tests to add the new fields.

These tests will have a much lower maintenance cost when adding fields
like that; we just need to run tests with PUSHSOURCE_UPDATE_BASELINE=1,
check the diff of the YAML files and commit them if the update looks
correct.