Fix missing deps and do manual pip-compile
Dependencies in CI haven't been updated in this repo for a while
because dependabot broke (unable to install rpm-py-installer).
Fixing this is a goal, but for now do a one-time manual update.

Additionally, some deps were missing from test-requirements.in
(pytest-cov, pylint). Fix them. Based on the comment in
test-requirements.txt, it seems like the file wasn't properly generated
from the .in file the last time it was touched.