tests: extend koji dump util with more fields
Having progressed with support for container images, I've learned
exactly which fields are needed on koji builds. Update the dump
utility to support them all.

Also make it support an 'all' NVR which redumps all existing builds,
so when the format changes like this it's possible to regenerate
them all easily.