Introduce ContainerImagePushItem.dest_signing_key
This field, filled in by ET, contains the signing key which ought
to be used to generate signatures when the image is pushed.

There is one limitation: we only support a single key here even though
the ET response theoretically supports more than one. This was done
because there are currently no use-cases for multiple keys, and because
none of the existing push code in Pub supports it (despite some buggy
attempts to do so).