koji: tolerate container builds with empty 'image'
A minor tweak to tolerate cases of a container build which has an
'image' field present but an empty dict. Rather than refusing to
handle it, let's continue and generate near-empty push items.

This is probably irrelevant to any real builds produced in recent
years. The motivation is improved compatibility with Pub, which
does accept builds like this and has some tests based on it
(though it's unclear if any such builds actually exist beyond
Pub's testdata)