koji: avoid some duplicate logs
In some error cases, a pattern was being used here like this:

- make an error message
- log the message as an ERROR
- raise exception with same message

The idea behind it is that, when raising the exception, we don't know
if anyone's going to catch it and log appropriately. If they don't,
then we may have created a situation we can't debug, and we'll surely
wish we had added that log message.

However, if the exception *is* caught and logged then this ends up with
duplicate messages.

Based on testing with Pub, the "duplicate message" case is more common,
so let's stop doing that and just raise the exception without logging an
error. We can still log as DEBUG, which will not show the message by
default but still gives us a way to enable the messages if needed.