Add enter/exit support to fix lifecycle/resource leak issue
Previously, sources were used like this:

    source = Source.get(...)
    for item in source:
        ...

Problem: a lot of the sources spawn additional threads internally and
there was no way to clean them up other than relying on GC (which
is problematic if there can be cycles between threads).

This was a bit of a design flub which was missed since the typical
use-case for a source is to only create one of them in the lifecycle
of your task/script/etc. However, it becomes relevant in cases such
as test suites, where hundreds of tests might each create Source
instances - leading to hundreds of threads spawned and never cleaned
up until all tests finish.

To fix this we now provide support for with-statements, which really
should have been there since the beginning. Now the recommended way
to use a source is:

    with Source.get(...) as source:
        for item in source:
            ...

It is not technically mandatory to use it this way, so only a subset of
tests were updated to use with statements. Most notably the baseline
test was updated to verify that there are no thread leaks when a with
statement has been used.

A related note in docs about being able to iterate over sources more
than once has been removed. In practice, none of the sources were
designed for it and none are being used in that way, and the addition
of with statements makes the situation even less clear / less likely
to work, so just drop that for now.