Fix missing dependency 'scandir' on older python
We try to use os.scandir, but fall back to scandir.scandir on
python < 3.5; this dependency wasn't declared, and it should be.

Previously missed because, on those python versions, scandir tends to
be pulled in via pytest => pathlib during tests. But this is not true
for all versions of pytest.