Mitigate spurious TimeoutError for slow caller [RHELDST-6064]
Previously iterating over any *Source object suffered from
spurious TimeoutError when caller was slow. The generator 'as_completed'
is unable to reset or stop timer for timeout, the timer for timeout
starts in original call to this generator.

Instead an alternative generator called 'as_completed_with_timeout_reset'
was implemented in helpers.py. It provides following functionality:
- yields finished futures as soon as any of them is finished
- after a finished future is yielded, there is no timeout check in
  place after yield
- timeout is only applied when next item from generator is requested

This approach allows caller to work as long as it needs, TimeoutError
is raised only in case when next requested item from pushsource
takes too long to yield to caller.

Fixed a couple of imports in order to pass pylint checks.