registry: fix list handling for some arguments
The point of the list_argument helper is that it will accept both
a true list (preferred in the case of programmatically creating a
Source) and a comma-separated string (preferred in the case of
obtaining source by URL). Calling split() on the input defeats the
purpose.

Fix it to use list_argument appropriately and ensure that each of
the 'list' and 'comma-separated' cases are covered by at least one
test.