ci: bump py3 version to 3.9
Use python 3.9 as the new minimum version for py3.

The motivation for changing it now is to resolve the issue seen in
pip-compile PR #191. The problem is that pip-compile is already using
python 3.9 but other CI configs used 3.8. This is not valid, as the set
of resolved dependencies can differ between the two versions.

It could be solved by either downgrading the python used for pip-compile
or upgrading the python used elsewhere. In this commit I pick the latter
approach, seeing as python 3.9 is readily available on all the py3
target platforms.