Fix get_partial breaking special handling of 'url'
If a source is defined like this:

    class MySource:
      def __init__(self, url, arg1=..., arg2=...):
        ...

...and registered as 'my-source', then as a convenience one may
refer to the source as 'my-source:/some/url' rather than
'my-source:url=/some/url'.

Problem: if get_partial was used to wrap the source, this special
handling for the url argument would be unexpectedly lost.

Fix it by ensuring that get_partial propagates whether or not a
source accepts a 'url' argument. Since functools.wraps does this
natively on python 3.4+, it will be possible to drop this when py2
support goes away.