errata: improve error handling on modulemd source lookup
If we're processing modules in ftp_paths and we fail to find an existing
ModulemdSourcePushItem for a given entry, we would previously give an
error message indicating that modulemd sources are missing from koji.

Actually, that is only one of two reasons we might fall into this error
case. The other reason is that the same modules might have been missing
from the cdn_file_list response.

Update the code to differentiate between the two cases. In the ET case,
we will not currently treat that as a fatal error since the behavior of
these methods is not strictly defined. However, it's sufficiently
unexpected to generate a warning.