Add stub files for all model classes except those for retrofitting attrs
There are a number of model class which are just "empty" subclasses of other models (e.g. ProductIdPushItem(PushItem)), which don't define any additional properties or methods outside of those provided by their base class. It may not be necessary (not 100% sure yet) to stub such classes, but it does provide a place where future stubs will go when needed.