Introduce FilePushItem.display_order
This attribute provides a way to manipulate the order in which files are
displayed when presented in a UI.

The field can be populated in staging metadata, at the same level of the
existing 'version' field. This format is the same as Pub apparently
supported prior to migrating to the pushsource library.

The absence of this field from pushsource up to now could be considered
an oversight. Users have managed to get by without it due to some
lower-level tools able to manipulate the relevant Pulp fields directly,
but it would be preferable to fold this into the regular push workflow.