Refresh integration with black, drop from 'tox -e static'
black was previously integrated into both tox and pre-commit.
The version used by tox would be managed by pip-compile.

This causes some issues: if our pip-compile github action updates
to a new version of black which adjusts any formatting, this will cause
CI to fail as it does not automatically run black over all files and
include changes in the commit.

pre-commit.ci on the other hand does support doing exactly that.
Therefore, for black and similar "auto-fixing" tools, it seems better to
leave them solely to pre-commit and not try to include them in tox.

As part of this change, the pre-commit config for black was updated to
ensure pre-commit autoupdates will succeed from this point onwards.