Use slots-based classes for pushsource models
Some of the commands based on this library are having memory usage
issues. According to my testing of a simple script which loads ~12000
push items, switching to slots-based classes will reduce the memory
usage of pushsource objects by 90%, and the overall memory usage of the
test script by ~33%. Let's do that to relieve the memory pressure.