Add support for "RAW" content via DirectoryPushItem [RHELDST-10120]
DirectoryPushItem is a PushItem where "src" is the path to a "RAW"
directory, and the "dest" is the directory's name under the staging
area.

For example, provided the following staging directory,

/
└── staging
    └── rhel-atomic-host-ostree
        └── RAW
            ├── a-file
            └── a-dir
                └── some-file.txt

If pushsource is asked to load "staged:/staging", it will generate
a DirectoryPushItem(src="/staging/rhel-atomic-host-tree/RAW",
dest=["rhel-atomic-host-ostree"]). It will not generate any
FilePushItem objects for the files inside the directory.