Work around python2 strptime thread-safety bug
Astonishingly in python2 there is a crash bug in strptime which can
occur if the very first call to the function happens concurrently from
multiple threads. This is https://bugs.python.org/issue7980 which was
closed as "Won't Fix" since the problem is resolved in python3.

It can be worked around by calling the function once at import time,
when we're confident there won't be concurrent calls. Workaround is
inspired by [1].

[1] https://github.com/boto/boto/pull/1940/files