Allow locating pushsource backends via entry points
This library supports custom backends. A problem can arise: the
backend has to be registered by a call to Source.register_backend, and
if implementing a backend in a library, there may not be any place to
put that call where it's guaranteed to happen by the time it's needed.

Fix this by adding a 'pushsource' entry point group. A library can
define an entry point pointing at a module which calls register_backend.
If so, pushsource library will ensure the module is imported during the
first call to Source.get().