Promote examples/list-push-items to a real command
This example script dumps whatever's found when given a particular
pushsource URL. I'd like to promote it to a proper command,
pushsource-ls, which is installed by pip. It seems to be quite useful as
a development tool and also in demonstrating this library's concepts.

Compared to the example script, some minor changes have been made:

- a simple config file is introduced; hardcoded URLs were OK for an
  example, but it should be properly configurable for a real command

- don't default to black format since not everyone will have that
  installed

- tweak it to be python2-compatible