Add caching for a few fields which often have duplicates [RHELDST-11377]
For push item sources such as staged and errata, fields like
'signing_key' and 'dest' will often have the same value across most or
even all of the loaded objects. This leads to tens of thousands of
copies of the same strings, wasting a bit of memory.

We can achieve some modest savings by caching/sharing values across
items. It is safe, since the shared values are immutable. This commit
implements such sharing for a handful of fields.

A few different caching approaches were tried here, but for most of
them the overhead of the cache exceeded the memory saved. In the end the
chosen strategy was to cache only the last couple of values per field,
which seems odd but works due to the typical usage patterns of push
items (constructing a lot of similar items at once). An included doc
string has some actual numbers on the savings.