Fix multiple levels of 'src' existence polling
297ad8bdbac33d760 added polling for the existence of files in the 'src'
attribute before yielding items.

However, the way it was implemented would tend to cause multiple
levels of polling because sources could be wrapped with a SourceWrapper
more than once. This is a common case in practice as several sources
used by Pub are built by wrapping the source underneath (e.g. 'brew'
source is just a 'koji' source with some arguments prefilled, and both
layers would end up with a SourceWrapper, both of which would run the
src polling logic).

Fix it up by a minor tweak to avoid creating multiple layers of
SourceWrapper.