Also poll for koji paths
Koji source checks if source paths exist and raises and exception if
they don't. Add a polling mechanism for this part of the code,
because otherwise the push items will fail to be created before the main
polling mechanism is even attempted.

Also fix a bug where polling is performed 1 less time then expected.