Only poll for highest priority signing key
Signing keys are added in the order of priority, and we want to use the
highest priority one that's present. Thus, we have to poll until the
highest priority key appears, or until we are sure that it's simply
missing. If the key fails to appear during the polling period, try the
remaining keys with decreasing order of priority. No more polling is
done, since the keys should have appeared during the first polling.