Make pushsource compatible with attrs >=22.1.0
In attrs version 22.1.0 the deprecated 'cmp' attribute was removed. This
has broken the dependency for pushsource, because the attributes are no
longer correctly initialized. AttrsRenamer checks whether a required
__init__ parameter is present in the created Attribute, and constructs a
new Attribute with only the present parameters. This is done with the
assumption that some constructor parameters will be removed in the
future, and this check would prevent the object construction from
breaking. In the attrs version 22.1.0 the deprecated "cmp" attribute was
removed and so it is no longer passed as an argument to the constructor.
This causes an error, since this argument is still expected in the
constructor, even though it doesn't do anything [1]. So let's also check
if the argument is in the constructor signature, and if it is, set it as
None. The method inspect.signature is only available in Python 3. This
is not an issue, since the attrs version that breaks pushsource (22.1.0)
also dropped Python 2.7 support, meaning that the breaking behavior may
only happen on Python 3.

[1] https://github.com/python-attrs/attrs/blob/main/src/attr/_make.py#L2504