Only poll for existence if item source is a file
Recently, a polling mechanism was introducted to pushsource to check if
a file exists before continueing the push. This was necessary because
mounted volumes (where the files are located) are snapmirrored and
there's a time delay before new files are propagated to the snapmirror
volumes. There is an unwanted side effect where all push items are
checked, even those that don't contain a file in their source
(namely container images that contain the image URL instead).
This is undesirable behavior, because image URL is not a file
and thus will never be found on disk. This bug significantly slows the
pushes, since the items will be polled until timeout is reached. Fix it
by only checking push item sources that start with "/", indicating
that they are local files.