Allow KojiSource to fetch VM Image PushItems
This commit changes the `KojiSource` to allow fetching Virtual Machine
Image items from Koji.

It uses the currently available metadata from `Koji` in order to build
the expected type of `VMIPushItem`.

As some information may be missing in Koji it's expected to the consumer
of these PushItems to evolve them by adding the remaining info from some
external service.

Refers to EXDSP-1781