Poll for staged content
Staged content is checked before the main polling mechanism, meaning
that PushStaged code fails unless the staged content is present in the
snapmirror. Since it's possible that staged content may be only
partially copied to the snapmirror, maximum snapmirror delay time must
be waited to ensure that data is up to date.

There are three scenarios:
- Staging directory doesn't yet exist in the snapmirror. Wait the full
  copy delay period.
- Staging directory exists, but its mtime is recent (what "recent"
  means can be modified with an environment variable). This indicates
  that changes to the directory were made not long ago, requiring the
  wait to ensure up-to-dateness.
- Staging directory exists and its mtime is older. Skip the wait since
  it was likely last modified a while ago.

This implementation assumes that push-staged is ran with only one path.
Otherwise, the code will wait the full poll_timeout for each path, which
is unnecessary and time inefficient. Running push-staged with multiple
paths is apparently not a real use-case, so it should be fine in 99% of
cases.