Fix bad validator on VHDPushItem.legacy_sku_id
This commit fixes a bad behaviro on `VHDPushItem.legacy_sku_id' which
should only be set whenever `support_legacy` is True.

If the constraint is respected the value of `legacy_sku_id` can be any
string, however the previous version was limiting it in `V1` or `V2`
(same as `generation`) which shouldn't be the case.