Add public_image flag for AMI images [RHELDST-16452]
The `public_image` flag designates whether the AMI image is allowed to
be released publicly (shared with group "all") as some images
are expected to be "shipped live" by going through all the standard
steps with the exception of being shared publicly.

This commit also fixes logic for getting the `ena_support` flag.
Previously, when the flag was set to `False`, the
`AmiPushItem._from_data()` method unintentionally converted it
to `None`. Now it remains set to `False`.