Fix CI failure installing rpm-py-installer
This became broken by the release of pip 23.1 and virtualenv 20.21.1.
Details on bug report:
https://github.com/junaruga/rpm-py-installer/issues/276

Until there is a proper resolution, pin to an older virtualenv to avoid
the issue.

Since the relevant deps here are used during installation of tox itself,
the pinning unfortunately cannot be done using the requirements files
consumed from *within* tox.