Remove setting src in container push items
Since container pushes were switched to Quay, this field is no longer
used. However, the file's presence in the mounted volumes is still being
polled before the push can begin. This can cause the pushes to take much
longer than necessary. Remove the field so that the unnecessary wait is
skipped.