Allow KojiSource to load the "boot_mode" from Koji
This commit allows `KojiSource` to read and set the VMI attribute
`boot_mode` by reading the boot_mode from Koji, which is being set by
`OSBuild` in `extra` for each VMI archive.

The `boot_mode` is an optional Enum named `BootMode` which is defined in
`vmis` module as well.

Refers to EXDSP-2127